3mul

+++ 0x0A-argc_argv/3-mul.c
@@ -17,7 +17,7 @@ printf("Error");
 return (1);
 }
 
-printf("%d\n", argv[1] * argv[2]);
+printf("%d\n", (int)argv[1] * (int)argv[2]);
 
 return (0);
 }
